+++ src/App.js
@@ -1,6 +1,6 @@
 import { Route, Routes } from 'react-router-dom';
 import NavBar from './components/NavBar';
-import BookList from './components/BookList';
+import BookPage from './components/BookPage';
 import Category from './components/Category';
 
 import './App.css';
@@ -10,7 +10,7 @@ function App() {
     <>
       <NavBar />
       <Routes>
-        <Route path="/" element={<BookList />} />
+        <Route path="/" element={<BookPage />} />
         <Route path="categories" element={<Category />} />
       </Routes>
     </>
