+++ src/components/BookPage.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import Book from './BookList';
+import Bookform from './BookForm';
+
+const BookPage = () => (
+  <div>
+    <Book />
+    <Bookform />
+  </div>
+);
+export default BookPage;
